shader practices answer3
정점이 -0.5로 시작하는 부분이 있는데

fragment shader에서는 음수가 clamp되어서 0이됨

그래서 검정색으로 나옴.

Think about this for a second: the output of our fragment's color is equal to the (interpolated) coordinate of
the triangle. What is the coordinate of the bottom-left point of our triangle? This is (-0.5f, -0.5f, 0.0f). Since the
xy values are negative they are clamped to a value of 0.0f. This happens all the way to the center sides of the
triangle since from that point on the values will be interpolated positively again. Values of 0.0f are of course black
and that explains the black side of the triangle.